requirements.txt
accelerate==1.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
airportsdata==20250706
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
beautifulsoup4==4.13.4
bert-score==0.3.13
bitsandbytes==0.42.0
bleach==6.2.0
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.21.1
cloudpickle==3.1.1
cmake==4.0.3
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.6
cssselect==1.3.0
cssutils==2.11.1
csvw==3.5.1
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
datasets==3.6.0
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dlib==20.0.0
dlinfo==2.0.0
dnspython==2.7.0
easydict==1.13
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.13
fastjsonschema==2.21.1
fastrtc==0.0.30
fastrtc-moonshine-onnx==20241016
ffmpy==0.6.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
genson==1.3.0
google-crc32c==1.7.1
gradio==5.43.1
gradio_client==1.12.1
greenlet==3.2.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
insightface==0.7.3
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
iso3166==2.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpath-ng==1.7.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
kernels==0.9.0
kiwisolver==1.4.8
kokoro-onnx==0.4.9
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mlx==0.27.1
mlx-lm==0.26.3
mlx-metal==0.27.1
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.1
multiprocess==0.70.16
murmurhash==1.0.13
narwhals==2.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.1
openai==1.100.2
openai-whisper==20250625
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.2
outlines==1.2.3
outlines_core==0.2.11
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
phonemizer-fork==3.3.2
pillow==11.3.0
platformdirs==4.3.8
playwright==1.53.0
plotly==6.2.0
ply==3.11
pooch==1.8.2
preshed==3.0.10
prettytable==3.16.0
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3986==1.5.0
rich==14.0.0
rouge==1.0.1
rpds-py==0.26.0
ruff==0.12.10
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
segments==2.3.0
semantic-version==2.10.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
simsimd==6.5.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
stringzilla==3.12.5
supervision==0.26.0
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.9.0
timm==1.0.19
tinycss2==1.4.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.8.0
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.168
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
webrtcvad==2.0.10
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
